Update for traffic signals class and others
I intend to add a branch for traffic signal sometime later on today.  Above are the additions to the includes dot h that I think will be needed to range bound some of the variables.  These additions should also obviously be helpful in some of the other functions as well.