Add files via upload
Changes found here:
http://pastebin.com/jZ3PaUSi

It turns out that constant data members cannot be initialized in subclasses if they're initialized already by the constructor in the super class. We may be able get around this in the future by making TransportMode and Vehicle virtual, but for now I just made the new "activeEmergency" bool a normal data member (not a constant).

I also added a debugging preprocessor directive so we easily toggle whether we want to have output in the constructors, found in Includes.h (but used in the all the vehicle .cpp files).