Traffic Signal Class
Hi guys.  I have added a traffic signal class file for our team to work on and complete.  The file is very raw and even lacks constructors so it will not run.  I have uploaded it mostly so you can see where we are and begin the process of rounding it out as a team.  Feel free to write any methods you think are germane.   Perhaps we should meet next week to discuss.   Spence.