Comments to Traffic signals.
I like the simple test loops you have designed. If information is unavailable then error is displayed. 
I like the test bool of light or stop sign. I think I will steal that test loop.
I am not sure why on line 56 we test for number of directions. I think we should assume all intersections are 4 ways. Any 3 way or 5 way will have a specific class to call for the intersection.  Or am I missing something there?
I will add a directionNumber to my code for each of the directions. I also want to use the setters you built for latitude and longitude, are we using a location for the map? Are we still going to number the intersections? 
I will focus on putting together the getters and setters for the class. Get the generic code together to be called at any moment.