Road - Kyle Cilley
So far it just inputs using the DataReader into an Array of Objects(Roads) to set different data members for each object/Road in the array.
The Loop in main sets Street Name then the Amount of each type of lane in each road. It also takes the amount of each type of lane in each road and multiplies it by each type's width and adds it up and sets it as that road's width. 

I used random streets and values to determine if it was inputting correctly. Until a spreadsheet with actual data is created.

Working on next
- Needs actual Lane widths, Amount of Lanes and Latitude/Longitude Coordinates attached to each road.