Add files via upload
These three files constitute a first stab at a program to read values from the link data spreadsheet using the Data Reader and place them in an array.  It does not run because the interface with the data reader has not yet been discovered by the programmer.  I hope to remedy that this weekend.  Hopefully, this array will contain data useful specifically to the Roads group.  Any other group may also use it.