Notes on this version of Main from in lab 11-19-16
/***************************************************************************************
David Sant
Spence Wilson
Isaac Haas
Last submitted on: 10/22/16
last update: 11/19/16
Uses: all of the files

This is a real clean version of main that removes everything not related to what I was adding.  

Added a intersectionsDataSTUB.H and .cpp file that take in the array of MapOBJs add put 
intersections on the array.  

(these STUB files will be replaced with real files by the data team)

Added a function to main that puts roads between the intersections as well as 
	to the edge of the array

Modified MapOBJ to add a constructor 'map[r][x] = new MapOBJ(true)' that accepts a bool
	value.  This bool makes road1 true.

Added a function that outputs the resutls

***************************************************************************************/